fix: some solidity code and new deployment

+++ src/IStudentNft.sol
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.21;
 
-import "openzeppelin-contracts/contracts/token/ERC721/IERC721.sol";
+import {IERC721} from "openzeppelin-contracts/contracts/token/ERC721/IERC721.sol";
 
 interface IStudentNft is IERC721 {
-	function mint(uint tokenIdToMint) external;
+	function mint(uint256 tokenIdToMint) external;
 
-	function burn(uint tokenIdToBurn) external;
+	function burn(uint256 tokenIdToBurn) external;
 }
